Extract dashboard into a class.
In order to make the dashboard testable it's useful for it to be its own
class.

Create a /lib directory.
Move the guts of the old dashboard file into a Dashboard class.
Move the Game class into /lib.
Create a new dashboard.rb that performs identically to the old one but
now uses the Dashboard class from /lib.